move range to its own module and some clean up
Get rid of non-Median BVH build.

Clean up printfs.
